idk what im doing
Started working on Mazesolver, but mostly just tried to understand instructions, mainly the idea of a worklist. Added needed methods to Square in order to see if the Square had been called already or not. Not done working, just gonna take a short break and didn't want to leave unsaved work.